Move map_nested tests to primitive package
